Revert "Merge branch '3.7/media' into android-3.7"
This reverts commit cccaca837f5f65d30131bde2fa39f882680b2c75, reversing
changes made to a43c50d08b4d86278220537d458e72c11a2d01ed.

+++ drivers/media/tuners/xc5000.h
@@ -35,7 +35,7 @@ struct xc5000_config {
 	u8   i2c_address;
 	u32  if_khz;
 	u8   radio_input;
-	u16  xtal_khz;
+	u32  xtal_khz;
 
 	int chip_id;
 };
